Support multipart/form-data PUT requests to API
I had to patch ``ModelResource.put_detail`` to avoid accessing
``request.body`` (we don't need it anyway) and triggering the
"You cannot access body after reading from request's data stream"
error when the request's format is multipart/form-data.
http://stackoverflow.com/a/12576374/386210  breaks down
what's happening here.

I also had to implement ``AssetResource.hydrate`` to convert empty datetime
fields from empty strings to ``None``. When sending a request serialized
as JSON, the date fields are sent as ``null`` which gets converted to
``None``, but when the requests are sent as multipart/form-data the empty
date fields are empty strings, which causes problems when trying to convert
them to ``datetime`` objects.

Finally I added a note and a copy of Tastypie's license explaining code
replicated from Tastypie and the reasons for this.

Addresses #738