Support asset image replacement in IE9
In IE9 (and lower) we have to send data in a form inside a hidden iframe
element. The HTML spec doesn't support using the PUT method in a form, so
we need to implement ``AssetResource.post_detail`` so the server will
accept the request.

Since this breaks the semantics, at least as Tastypie defines them, of a
POST to the detail endpoint, we should only implement this if the ``iframe``
querystring parameter is present.

I refactored some of the code in ``AssetResourceTest`` to avoid repeating
code in the tests for this fix.

Closes #749, addresses #738