Don't re-fetch section's assets
``SectionEditView`` defers rendering of the subviews for each
container until the assets are fetched.  On its initial render, it
fetches the assets.  However, this had the side-effect of re-fetching
the assets when switching between workflow steps.

This is unneccesary and actually causes problems with the model event
bindings in ``SectionAssetEditView`` since the model instance in
``SectionEditView.assets`` is no longer the same as
``SectionAssetEditView.model``.  One way this issue showed up was
in #764, where ``SectionAssetEditView`` instances did not re-render
when an asset was removed from a container by clicking on the remove icon.