Set flag for already fetched datasets
Set the ``_collectionFetched`` flag to ``true`` for datasets that have
already had their ``datasets`` property initialized. This is most often
the case when an asset has been removed from a section container and
then re-added by dragging it from the asset drawer.

This change fixes a bug where the data set list was not being rendered
after an asset with datasets was dragged from the asset drawer after it
had been removed by changing the section layout.

Addresses #583