Add "save and add another" button to data set form
Add a separate "Save and add another" button to the form for adding
datasets.  Clicking the "Save and add another" button resets the
form, but keeps it visible. Clicking "Save" hides the form.

In order to implement this, I had to break rendering of the data set
list into a separate method.

I also had to add logic to reset the add data set form. It seemed like
the cleanest way to do this was to destroy the old form view and create
a new one and re-render. It would be nice of ``Backbone.Form`` instances
had a reset method.

Addresses #583