Switch to asset display state after adding data set
Switch back to the display state after a new data set is added
instead of returning to the data set list view.  This is achieved
by proxying the ``create:dataset`` event along with its
``postSaveAction`` from ``AssetDataSetAddView`` through
``AssetDataSetListView`` and on to ``SectionAssetEditView``, which
will switch to the ``display`` state if ``postSaveAction`` is
``close``.

Before switching to the display state, refresh the asset model from
the server in order to display the most recent version of the
rendered data set list (in the asset model's ``content`` attribute).

Addresses #583