Mitigate flash of content after adding data set
A flash of content occured because ``AssetDataSetListView`` toggled
the display of the data set list and the add data set form before the
state of ``SectionAssetEditView`` was updated and the view re-rendered
while waiting for the asset model to sync with the server. This is
creating a ``sync`` state of ``AssetDataSetListView`` and setting this
state before calling ``fetch`` on the asset model.

Addresses #583