Support editing asset datasets in builder
``DataSetResource`` had to be updated to support the ``put_detail``
method. I also had to update the ``hydrate_file`` method to handle
updating an existing ``LocalDataSet`` instance (without updating
the file field)

Also, the ``DataSet`` Backbone model's ``schema`` method had to be
updated to handle selecting only one of the ``file`` and ``url``
fields in the case of an existing model.

I factored common functionality shared between
``AssetDataSetAddView`` and ``AssetDataSetEditView`` into a base
view called ``DataSetFormView``.

I changed the text of the reset button in ``AssetDataSetListView``
to be "Close" instead of "Cancel".

I also updated the propagation of events from
``AssetDataSetListView`` from to ``SectionAssetEditView``. Instead
of propagating the ``save:dataset`` and ``create:dataset`` events,
I keep track of when datasets have been added or changed and
trigger a ``close`` event on ``AssetDataSetListView``. This event
has a ``refresh`` argument that is truthy if the parent view
should fetch and re-render the content.

Addresses #583