Move asset dataset list markup into a template
Refactored ``Asset.dataset_html`` so that the markup is in a template.
This is much cleaner and more extensible than concatenating a series of
strings.

I also updated the markup so that the data set title is no longer a link
and the download icon follows the title instead of preceding it.

Addresses #583