Automatically set links_to_file on save
Set ``DataSet.links_to_file`` automatically when saving an
``ExternalDataSet`` model.  ``mimetypes.guess_types`` is used to
guess whether or not the URL stored in ``ExternalDataSet.url``
represents a downloadable file.

This implementation prohibits the manual overriding of this
flag in the Django admin, which is problematic for downloadable URLs
that don't have some kind of filename in the path, but this seems
like an edge case that doesn't warrant additional complexity.

Addresses #583