Open data set links in a new window
Open data set links in a new window using ``window.open``. This
is accomplished by giving the links an ``external`` class and
wiring up the correct click handlers in the Backbone views' ``events``
property.

Since this is a behavior that is used throughout the app, I broke
the click handler out into a function in ``storybase.utils`` and
updated the code to use this function where I could.

Addresses #583