Indicate which data set form fields are required
There isn't a great way to indicate that either one field or another
is required (in this case the url or file fields) until #767 gets
implemented, so I'm using fieldsets and legends to hack around this
for now.

I added some new utility functions to storybase.forms to accomodate
setting classes, attributes and label text for required fields.
Eventually, we can make use of these for other form instances.

Addresses #583