Properly validate data set form on creation
Fix the lack of validation when the user provides both a file and
url field when creating a new Data Set as reported at
https://github.com/PitonFoundation/atlas/issues/583#issuecomment-19067476

We use DataSet.validate to detect that one and only one of the url
and file fields as provided.  However, model validation doesn't get
called unless a model instance is passed to the form constructor.
So, if a model instance isn't passed to the parent view, construct
an empty DataSet model.

I also decided to report these errors with the field instead of
at the top of the form.

There was a bug in ``Backbone.Form.validate()`` where field errors
reported by model validation were not added to the return value. I
decided to upgrade the version of Backbone Forms that we use to the
current stable version (0.12.0) as this includes a fix for the bug.

Our styles are based on the old Backbone Forms template markup, so
I overrode the templates with the old ones.  Eventually, we should
update the styles to reflect the current templates.

Addresses #583