Display asset datasets on story detail page
Add a ``asset_datasets`` method to the ``Story`` model to get a
queryset of only the datasets that are tied to assets that are
visible in the story. This allows ``Story.datasets`` to hold a list
of all datasets that have been associated with a story if we want
to also support datasets not associated with assets or the ability
to re-add a dataset to an asset.

Use Bootstrap's Dropdown Menu component to display the list of
datasets without taking up too much screen space on the story
detail page.

Addresses #733