Update asset form labels in schema argument
After the upgrade of Backbone Forms in
f9c09d7ed4e4003b26c1ab4acbf422b76d14130f, the strategy of altering
asset and featured asset form labels by updating properties of
``this.form.schema`` in an ``updateFormLabels`` method didn't seem
to work. It seems to make more sense to take the approach used with
the ``DataSet`` forms and to update the schema before instantiating
the form.

Addresses #583