Fix editing of uploaded image assets in builder
Allow editing asset types that are backed by the ``LocalImageAsset``
Django model when no new image file is specified.

This is particularly important when editing chart assets, since
one might want to update the asset after attribution after the
model has been created.

Update ``AssetResource.get_object_class`` to return the model class
of ``bundle.obj`` when updating an existing model. This returns the
correct class, even when an identifying bundle field is missing,
avoiding the "You must specify an image, body, or url" error when
updating an existing model instance, but not changing the ``image``
field.

On the front end, remove undefined/empty ``image`` attributes
before updating the Backbone ``Asset`` model so that the existing
value of the ``image`` field will not be clobbered by a null value.

Addresses #794