Copy template properties as placeholders, not values
When initializing Story and Section models from templates in the
builder, store the value as a suggestion rather than setting the
field value.  This allows the suggested value to be displayed as
a placeholder in the various fields instead of as the form field's
value.

The suggested values from the template are stored as Backbone
model attributes with "_suggestion" appended to the normal field
name, e.g. "summary_suggestion".  The Story fields that receive
suggestions from the template are summary and call_to_action.
The Section fields that receive suggestions from the template are
title.

The copying of suggested values happens initially in Story.fromTemplate
and in Story.suggestionsFromTemplate for already-saved stories.

Addresses #722