Switch from container help to section help
Switch from container help to section help when drawer is closed.
Previously, there was no way to go back to a section's help if
an asset container's help had been selected.

In order to implement this functionality, I added an options
hash to ``HelpView.show`` and ``HelpView.set`` which supports
a ``remember`` key that causes the previous help value to be saved
when it is replaced.  Then, ``HelpView.pop`` can be called to
retrieve the last-saved help value.  ``HelpView.pop`` is called
when a ``do:pop:help`` event is triggered over the event bus.

Addresses #641