Searchable help
Create a ``SearchIndex`` subclass for the ``Help`` model to include help
in the site search.

However, not all help should be indexed since some
of it only makes sense in context (e.g. inline help in the builder). Add a
``searchable`` flag field to the ``Help`` model to differentiate help that
should be included in the index.  In order to make this schema update,
the ``storybase_help`` app had to be initialized to be used with South.

Update the search results template to show the type of result (for now
either "Help" or "Story").

Add a detail view for help items so links in the search results go somewhere.

We want to recreate the existing help page, and allow the hand curation,
so the most straightforward way to do this is to create a custom Django
CMS Plugin that can be inserted in a CMS page region.

Addresses #507