Don't try to create notifications for authorless stories
While this case shouldn't occur in actual system use, in tests, the
``create_story`` API function is often called without an author specified.

The signal handlers used to create ``StoryNotification`` model instances
when a story is saved were failing when trying to check the
notification settings in the profile for the nonexistant story author.

This commit adds a guard statement to check that the story has an
author before proceeding with creating a notification.

Addresses #580