Pass include_host when testing method
As part of implementing #580, the API for
``Story.featured_asset_thumbnail_url`` changed so that the hostname
is included in the returned URL by default.  This breaks the
``storybase_story.tests.StoryModelTest.test_get_featured_asset_thumbnail_url_empty``
test.  Passing ``include_host=False`` to the call to
``featured_asset_thumbnail_url`` matches the old behavior that the test
is expecting.