Connect signals for real-time search index updates
The codebase had not been updated to use the signal processors pattern
(see
http://django-haystack.readthedocs.org/en/latest/signal_processors.html)
used to implement real-time search index updating in Haystack 2.0+.

This commit moves the connection and disconnection logic out of
the ``StoryIndex`` class and into a custom signal processor classs,
``storybase.search.signals.RealtimeSignalProcessor``. It also wires
up handlers to update the index for ``Help`` models.

The implementation is a little verbose, but I think that's ok in the
interest of clarifying what's being wired up.

Closes #817