Clean up asset/dataset form error handling
The code reflected a misunderstanding of how form validation and model
validation work together because the
[documentation](https://github.com/powmedia/backbone-forms#model-validation)
is confusing.

Backbone.Form's validation will call the model's ``validate()`` method.
If the model validation returns an object, it tries to set errors on
the form fields matching the keys of the error object. If a key doesn't
match, it's added to the ``_others`` array of errors in the form's
error object. So, ``_others`` will never include errors for form fields
and they should always be added to the form errors displayed above the
form.

I cleaned up the code a bit by moving variable declarations to the
top of the function definition, though the need to do that is
[highly debatable](http://stackoverflow.com/a/3685090/386210).

I got rid of an anonymous function declaration used in a
call to ``_.each()`` because it seemed inefficient to redefine the
function inside a for loop.

Addresses #767