Add views for stories of topics and tags
Widgets for lists of stories aggregated by different means
(organizations, places, projects, keywords (tags), topics) are created
by JavaScript that converts a URL for a page for that particular
taxonomy item into an iframe.

For instance the URL
http://floodlightproject.org/projects/finding-a-bite-food-access-in-the-childrens-corrid/
will be converted to a a widget listing stories for that project.

While views and URLs exist for taxonomy item pages for projects and
organization. They didn't yet exist for keywords (tags) and topics.

We've used the explore page as the means to list stories for topics,
places, projects and organizations, so I added a new view class
``ExplorerRedirectView`` which routes a slug-based URL to the explore
page with filters set. This provides the implementation for the
'/topics/<slug>/' path.

The explorer doesn't support keyword tags, so I made a simple view,
``StoryListView`` for listing stories for a given keyword tag.

Since StoryListView uses paging, I broke the paging styles into its own
less file, paging.less so and adjusted the class names slightly in the
search results templates so the pagination styles could be used in both
places.

Addresses #741