Update requirements for Django 1.5/Django CMS 2.4
Update the pip requirements file to reflect the required packages
for Django 1.5.* and Django CMS 2.4.*

Also, remove entries in the requirements file for packages that
are not directly required by this software.  They should get
installed automatically by pip when the explicitely named
dependencies are installed.

Remove the pip requirements file for Django 1.3 as we're no longer
supporting this version

Addresses #660