Export widget object as storybase.widgets
The widget script is now exported as ``storybase.widgets``. This
lets us export additional methods if we want to test them. It also
provides a way to check if the widget script has been previously
loaded and short-circuit the rest of the setup, eliminating the
previous requirement that the script only be included once.

The ``showWidgets()`` function now only processes placeholders if
the ``style.display`` property is not set to 'none'. This allows
``showWidgets()`` to be run multiple times, which is what happens
if the script is included multiple times.

Addresses #741