Makefile for building and compressing CSS, JavaScript
Use GNU make to compile LESS files into compressed CSS and
compress standalone JavaScript files using Google's Closure
Compiler.

To build and compress all LESS and JavaScript files run this
command in the top level of the directory:

    make

To build only the LESS files, run:

    make styles

To build only the JavaScript files, run:

    make scripts

There are even more atomic build targets.  Take a look at the
source of ``Makefile`` to see them.

Note that most JavaScript should be compressed using the
``django-compressor`` app, configured using template tags. The
rules in this makefile are for compressing stand-alone scripts,
such as the widget JavaScript file that are not loaded from a page
generated by a Django template.

As the makefile uses the ``lessc`` command, the ``buildless.js``
script is no longer needed.