Use language code in URLs
The use of ``i18n_patterns`` and ``LocaleMiddleware`` required
by Django CMS >= 2.4 causes paths without a langauge prefix
to no longer resolve.  ``LocaleMiddleware`` looks for the 404
and tries to redirect to the language-aware path, but this
isn't great for our tests that expect a non-302 response.

We could have passed ``follow=True`` to the test client
methods, but in some cases, we need to explicitely test that
the user is not redirected.

The easiest way to make sure we're testing correctly is to
just test the language-aware path.

Addresses #660