Use modified version of LocaleMiddleware
Use our own version of ``LocaleMiddleware`` because the Django version causes
requests to the REST API that return a 404 to be redirected to a non-existant
path.  For example, if the request to /api/0.1/stories/STORY_ID/
returns a 404 because the story with ``story_id`` of STORY_ID
doesn't exist, the middleware will instead redirect the user
to "/en/api/0.1/stories/STORY_ID/".  This path doesn't exist
as part of our API (which doesn't use language-aware URLs).

This is because of the way ``LocaleMiddleware.process_response()``
tries to automatically redirect from a non-language-aware path
to a language-aware path.

This is very similar to the problem described at
https://code.djangoproject.com/ticket/17734, but the fix doesn't
work because the URL configuration for Django CMS:

    url(r'^(?P<slug>[0-9A-Za-z-_.//]+)/$', details, name='pages-details-by-slug')

matches for language-aware paths of views that aren't
configured using i18n_patterns (in our case, the REST API).

Addresses #660