One widget view to rule them all
Move to a single widget view at the path /widget/ instead of a
separate view relative to each item.  This has the following
advantages:

* Nice error response, even if the user totally mangles the URL
* Ability to have a primary story + list view automatically get
  the primary story based on the list URL if a primary story
  isn't explicitely specified
* One (or two) URL conf entries, in one Django app, instead of
  multiple ones, spread across apps
* Ability to show a default widget if no query arguments are
  provided

Also, concerns about JavaScript size were overstated.  Twitter's
widget JavaScript is 26k.

As a result, I removed ``StoryListWidgetView`` and the related
test mixin.  Also, ``StoryListMixin`` no longer made sense, so
it was removed.

This change introduces one backwards-incompatible change. Since,
the widget view is no longer relative to the item detail views,
the widget JavaScript has to build the full URL of the view,
including the hostname.  The script assumes "floodlightproject.org".
In order to override this, the user must set the variable
``_sbBaseUrl`` in the page that includes the widget JavaScript.

For example:

    <script>
    _sbBaseUrl = "http://dev.floodlightproject.org/";
    </script>

Will cause widget iframes to point to
"http://dev.floodlightproject.org/widget/" instead of
"http://floodlightproject.org/"

Addresses #741