Fix "View Story" button URL for connected stories
Properly set the URL for connected stories for the "View Story" button
in the "Publish/Share" step of the builder.

To implement this, add a ``viewer_url`` method to the ``Story``
model, and a corresponding attribute to the response data of
``StoryResource``.

Also, refactor ``PublishedButtonsView`` to separate out the element
showing/hiding and rendering.  Also, re-render when the story's
``viewer_url`` attribute is changed.

Addresses #856