Handle connected story URLs in notification emails
Set the ``detail_url`` context variable to an empty string in the
case of a connected story and use the value of this variable
to hide certain content in the template that's related to a
story's detail page, which doesn't exist in the case of a
connected story.

Clean up indentation and simplify template markup in the
notification templates.

Fix whitespace in templates, especially the text templates.

Replace unicode apostrophe characters with ascii ones.

Default the greeting to say "Floodlight User" if an author's
first name hasn't been specified in their profile.

Add a ``builder_url`` variable to the template context

Addresses #856