Avoid hard-coding popup view paths in URL
To make the template code as polymorphic as possible, implement
``get_embed_popup_url()`` and ``get_share_popup_url()`` methods on
the ``Story``, ``Project`` and ``Organization`` models and implement
the ``get_share_popup_url()`` on the ``UserProfile`` model.

These methods call through to ``reverse()`` to look up the correct
view for the popups.

Addresses #824