Refactor embed/share button popup script
The purpose of these changes are to minimize hard-coding of
URL paths in JavaScript and to make button links point to a
fallback URL instead of "#".

The buttons for launching the embed or share popups used to have
an href attribute of "#".  Whenever possible, I think it's better
to have links that fall-back to somewhere.  So, I made
``EmbedView`` and ``ShareView`` that render the ``storybase/embed.html``
and ``storybase/share.html`` templates which simply wrap the
``storybase/embed_popup.html`` and ``storybase/share_popup.html``
templates in the ``base.html`` template.

Then, I changed the url paths for the embed and share popups from
``{{ OBJECT_DETAIL_URL }}/share-popup/`` to
``{{ OBJECT_DETAIL_URL }}/share/popup/``.  Note that
``{{ OBJECT_DETAIL }}/share/`` points to the standalone share page
view mentioned above.

This allows the ``jquery.storybaseshare`` plugin to default to
building the URL of popup content fetched asynchronously by just
appending 'popup/' to the ``href`` attribute of the link. This
value can still be overridden by specifying a ``popupUrl`` option
in the ``data-options`` attribute of the element.

This also eliminates the need for building the popup content URL
by replacing "id" in a template string. Instead of caching the
fetched popup content by slug/id, I use the URL as the key. Now,
the "data-story-id" option is no longer needed on the element
initialized by the plugin.

Add the ``jquery.storybaseshare`` script to the files handled by
``django-compressor`` to force users to download our updated
version of the script when we deploy.

Addresses #824