Default to shorter height for story-only widget
If there's only a single story in the embed widget, it looks
akward with the default height of 500px.  So, default to a height
of 240px if there's only a single story.  In order to detect the
case of the single story, factor out the placeholder URL parsing
from the widget URL generation functions.

Addresses #741