Upgrade Font Awesome to version 3.2.1
Add the ``icon-fixed-width`` class in the following contexts:

* dataset icons
* menu icons in the builder's WYSIHTML5 editor.
* connected stories icon in the "Latest Stories" list
* connected stories icon in the explorer

The icon styles used to apply to the ``:before`` pseudo-element
of an element with a ``icon-*`` class::

    [class^="icon-"]:before,
    [class*=" icon-"]:before {
      font-family: FontAwesome;
      font-weight: normal;
      font-style: normal;
      display: inline-block;
      text-decoration: inherit;
    }

and now they apply to the element with the icon class itself.
Restore the old styles for the ``:before`` pseudo-element for
elements with the ``icon-chevron-right`` and ``icon-chevron-down``
classes as we use these throughout the code as toggle icons. In
particular, this is eneded for the headings on the ``/help/`` page
and the "Find Out More" links in the builder template selection
step.

The headings on the ``/help/`` page (they have a class of
``faq-question``) had to have their styles updated to have have
``display: block`` (overriding the ``display: inline-block`` from
having an ``icon-*`` class).

The "divot" which uses Font Awesome to show the down arrow icon
when the user clicks the "Find Out More" links in the builder
template menu, had to be moved up by about 4px.

Similarly, the element with class ``guider_arrow`` used in the
builder tour popups needed to be moved up by about 8px.

Updated the toggle HTML menu icon in the editor to use the
``icon-code`` icon from Font Awesome instead of the string "HTML".

Addresses #868