Add link to story detail page from viewer
Add a link to the story detail page in the header of the story viewer.
This is particularly important for when the viewer is shown in a popup
from a widget.

Give the share link the class ``storybase-share-link`` so we don't
have to override the ``storybase-share-button`` styles and so it's
easier to match the styles of the other links in the ``#tools``
container.

Addresses #851