Fix summary layout on smaller displays
With a 50% column width, the table of contents overlapped the image.
Make the column with 48% to prevent this.

Tall featured images were also pushing the summary text outside of view.
Move the summary text to the top of the display, pushing the table of
contents down.

Add padding to the bottom or elements so they don't get cut off by
the footer.

Remove the byline as it's redundant with the omnipresent header bar.

Addresses #851