Refactor ViewerApp.initialize()
Add hooks to ``ViewerApp.initialize()`` to allow subclasses to
use it as their initialize method, or to call super without all
the repeated code.

Pass the router as an option to the view instead of the other way
around.  Instead of having the router call methods on the view,
listen for ``route`` events inside the view.

Addresses #873