Use ``attributes`` instead of custom option
Use ``attributes.class`` to set the class on the view's element
instead of ``options.elClass``.  Using attributes is more inline with
how Backbone sets elements on an element it creates.

Originally, ``elClass`` had been a property of the view class
instead of a key of ``options``.  Making it a key of ``options``
caused the options of the parent class to get clobbered, which broke
table of conents toggling.