Fix section URLs in viewer
After changing the use of the router so the views listen to
events on the router in e9feb772b815678e569a70dbdd31da5dbfa1e5d6,
we also need to make sure to call ``Backbone.history.start()``
only after the the main application view has been initialized.
Otherwise, the router event listeners haven't been wired up, and
the correct content won't be displayed when a path like
``/stories/whats-the-last-photo-on-your-phone/viewer/#sections/connected-stories``
is entered in the location bar of the browser.