Make activity guide interactions available anywhere
Originally, I had planned to just update the styles so that
activity plugins that weren't on a CMS page using the special
template would just show up as expanded.  However, their CSS
float property causes layout to break in some cases.  I decided
to make the JavaScript more universal so things would "just work"
instead of requiring editors to add classes or clearfix divs to
the page content.

Break JavaScript to expand and collapse the activity guide blocks
out of the template ``activity_guides.html`` into a jQuery plugin,
``jquery.storybase.activityguide.js``.

Update the plugin template
``cmsplugin_storybase/activity_plugin.html`` so it uses
``Modernizr.load()`` to asynchronously add the plugin file to
the page, and only download it once.

The plugin logic actually creates the container for the blocks
and applies the Masonry plugin to it, so there is no longer any
need for a custom  template.

Update the Makefile to compress the new jQuery plugin. Also add
some documentation that describes how to process the assets that
need explicit compilation or compression.

Addresses #874