Fix typo in ``{% url %}`` template tag
I had wrapped the argument to the url in quotes when I updated
the use of the ``{% url %}`` template tag to put the URL name
in quotes.  In this case, only the URL name and not the argument
should be quoted.  Quoting the argument caused a ``NoReverseMatch``
exception to be raised when trying to resolve the url.