Copy CMS page teasers on publish
Django CMS 2.4 switches to a moderation model where there are
two versions of a page, a draft and a published one.  Inside
``Page.publish()``, ``Page._copy_titles()`` is called to copy
the title models to the published page from the draft page. We
need to implement something like this for the teaser models too.
Luckily there is a published signal that we can attach to and call
a ``_copy_teasers()`` method that we patch onto the ``Page`` model.

Closes #875