First pass at auto-expanding text editor
The WYSIHTML5 editor makes it pretty hard to implement this. One of
the major challenges is that the editor copies the styles, including
height of the base textarea element, to initialize the rich text
iframe.  It then uses the copies of these styles to reset the iframe
styles whenever the editor loses/gains focus.  So, I had to read
through the source to figure out how to make sure the new height
we set isn't clobbered.

Updated ``RichTextEditorMixin.createEditor()`` to take an optional
``options`` argument that are passed on as options to
``wysihtml5.Editor()``.

Still to do:

* Make the auto-resize an option for all editors
* Get it to size down when the text decreases. This is easier
  said than done because there isn't a nice event to hook into
  and we'll probably have to listen to keypress events.

Addresses #844