Move editor growth initialization into createEditor()
Move the initializaiton of the event listeners on the
``wysihtml5.Editor`` instance to the
``RichTextEditorMixin.createEditor()`` method so that this
functionality can be used in places other than just text asset
bodies.

Also removed a hack that disabled the default event handlers from
the editor instance because it no longer seemed to be needed and
would have broken the auto-grow functionality.

Addresses #844