Removed percentage height from asset wrapper element
The wrapper div around the asset element had a percentage height
of 94% specified. In the side-by-side layout, when one of the
assets is a text asset, it caused the wrapper for the adjacent asset
to grow alongside the text asset as the editor auto-expanded. I can't
see any reason for the percentage height, and can't remember why we had
it, so I just removed this style.

Addresses #844