Convert Creative Commons endpoint data to JSON
Parsing the XML from the proxied Creative Commons endpoint, extracting
the HTML representation of the license and finally appending it to
the DOM was causing an error in IE8.

Instead, extract the important element on the server side and convert
the data to JSON, which is much easier to deal with in JavaScript.

Refactor ``ProxyView`` to have more hooks to make it easier to alter
the remote endpoint's response and (potentially) cache the response
in ways other than using the ``cache_page`` decorator.

Closes #878