First pass at text asset autosave
Save a text asset on the wysihtml5.Editor object's ``change`` event,
which is essentially fired as a result of a ``blur`` event on the
editor's element.

To reuse ``SectionAssetEditView.saveModel()``, add an optional
``changeState`` boolean argument to the method that can suppress
the changing of view state when the model is saved.

The behavior of the "Save" and "Cancel" buttons still needs to be
updated to take the autosave behavior into account.

Addresses #848